my-project/requirements.txt
 pytube 
 PyQt5 
 requests 
 openai-whisper
 fastapi
 uvicorn